refactor(vehicle): decompose edit_vehicle_screen via a form-state mixin + a presentational form — closes #3234
The screen half of #3234 (the entity half shipped in #3293). edit_vehicle_screen
was an 879-line god-screen: ~290 lines of UI-coupled ConsumerState logic (VIN
decode, adapter pair/persist, auto-population, calibration, save) sharing ~13
mutable form fields, plus a ~200-line build tree.

Split into three files, each below the 400 cap:

- edit_vehicle_screen_actions.dart (358) — `_VehicleEditActions`, a `part`
  mixin owning the 13 mutable form fields + the 13 imperative action methods.
  It's a `part` (not a separate library) so the fields/methods stay PRIVATE and
  every call site (`_type`, `_decodeVin`, …) is unchanged; the 2 infrastructure
  members it needs (`_ctrl`, `_formKey`) are abstract getters the State's fields
  satisfy implicitly.
- vehicle_edit_form.dart (281) — `VehicleEditForm`, a stateless ConsumerWidget
  rendering the PageScaffold + the section-card stack. The State passes the live
  form values + pre-built setState/action callbacks, so it's a pure view.
- edit_vehicle_screen.dart (879 → 308) — keeps only the load/dispose lifecycle,
  the prepop `ref.listen` refill, the discard `PopScope`, and `build`.

Pure relocation: identical behaviour, no API change. flutter analyze clean; the
full vehicle suite (748 tests, incl. all 48 edit_vehicle_screen cases — VIN,
calibration, adapter-in-place, prepop-race, persistence) stays green. Removed
edit_vehicle_screen from the file_length snapshot — closes #3234.
